+++ src/main/java/com/example/typing/entity/SingleResult.java
@@ -10,10 +10,10 @@ import java.time.LocalDateTime;
 public class SingleResult {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer id;
+    private Long id;
 
     @Column(name = "user_id")
-    private Integer userId; // FK: users.id
+    private Long userId; // FK: users.id
 
     private Integer score;
